better logging

+++ VM.c
@@ -108,8 +108,8 @@ void MAP_VM(void)
 
 void log_to_file(struct timespec start, struct timespec end, FILE* log_file, uint16_t op){
     clock_gettime(CLOCK_MONOTONIC_RAW, &end);
-    float delta_ns = (end.tv_nsec - start.tv_nsec) ;
-    fprintf(log_file, "op: %x, time elapsed: %f\n", op, delta_ns);
+    long long delta_ns = (end.tv_nsec - start.tv_nsec) ;
+    fprintf(log_file, "op: %x, time elapsed: %llu\n", op, delta_ns);
 }
 #else
 void MAP_VM(void){
